+++ vespa/src/test/java/ai/marqo/search/HybridSearcherTest.java
@@ -623,661 +623,4 @@ class HybridSearcherTest {
                             "lexical yql");
         }
     }
-
-    /**
-     * Test for sortBy feature in HybridSearcher.
-     */
-    @Nested
-    class sortByTest {
-
-        /**
-         * Test that verifies sorting of results based on a single sort field.
-         */
-        HitGroup helpGenerateHitGroupWithOnlySortFieldValue0() {
-            FeatureData f1 = mock(FeatureData.class);
-            when(f1.getDouble("sort_field_value_0")).thenReturn(-1e50);
-            Hit doc1 = new Hit("doc1", 0.55);
-            doc1.setField("matchfeatures", f1);
-
-            FeatureData f2 = mock(FeatureData.class);
-            when(f2.getDouble("sort_field_value_0")).thenReturn(1.0);
-            Hit doc2 = new Hit("doc2", 0.65);
-            doc2.setField("matchfeatures", f2);
-
-            FeatureData f3 = mock(FeatureData.class);
-            when(f3.getDouble("sort_field_value_0")).thenReturn(2.0);
-            Hit doc3 = new Hit("doc3", 0.75);
-            doc3.setField("matchfeatures", f3);
-
-            FeatureData f4 = mock(FeatureData.class);
-            when(f4.getDouble("sort_field_value_0")).thenReturn(2.0);
-            Hit doc4 = new Hit("doc4", 0.85);
-            doc4.setField("matchfeatures", f4);
-
-            FeatureData f5 = mock(FeatureData.class);
-            when(f5.getDouble("sort_field_value_0")).thenReturn(-1e50);
-            Hit doc5 = new Hit("doc5", 0.45);
-            doc5.setField("matchfeatures", f5);
-
-            FeatureData f6 = mock(FeatureData.class);
-            when(f6.getDouble("sort_field_value_0")).thenReturn(5.0);
-            Hit doc6 = new Hit("doc6", 0.95);
-            doc6.setField("matchfeatures", f6);
-
-            FeatureData f7 = mock(FeatureData.class);
-            when(f7.getDouble("sort_field_value_0")).thenReturn(6.0);
-            Hit doc7 = new Hit("doc7", 0.90);
-            doc7.setField("matchfeatures", f7);
-
-            FeatureData f8 = mock(FeatureData.class);
-            when(f8.getDouble("sort_field_value_0")).thenReturn(7.0);
-            Hit doc8 = new Hit("doc8", 0.80);
-            doc8.setField("matchfeatures", f8);
-
-            FeatureData f9 = mock(FeatureData.class);
-            when(f9.getDouble("sort_field_value_0")).thenReturn(8.0);
-            Hit doc9 = new Hit("doc9", 0.70);
-            doc9.setField("matchfeatures", f9);
-
-            FeatureData f10 = mock(FeatureData.class);
-            when(f10.getDouble("sort_field_value_0")).thenReturn(9.0);
-            Hit doc10 = new Hit("doc10", 1.00);
-            doc10.setField("matchfeatures", f10);
-
-            HitGroup hits = new HitGroup();
-            hits.add(doc1);
-            hits.add(doc2);
-            hits.add(doc3);
-            hits.add(doc4);
-            hits.add(doc5);
-            hits.add(doc6);
-            hits.add(doc7);
-            hits.add(doc8);
-            hits.add(doc9);
-            hits.add(doc10);
-            return hits;
-        }
-
-        /**
-         * Test that verifies sorting of results based on a single sort field with ascending order
-         * and last missing policy.
-         */
-        @Test
-        void sort1FieldWithAscOrderAndLastMissingPolicy() {
-            // build 10 distinct docs by hand
-
-            HitGroup hitsToSort = helpGenerateHitGroupWithOnlySortFieldValue0();
-            HybridSearcher searcher = new HybridSearcher();
-            String sortJson =
-                    "[{"
-                            + "\"field_name\":\"ignored\","
-                            + "\"order\":\"asc\","
-                            + "\"missing\":\"last\""
-                            + "}]";
-
-            // full-depth, no trim
-            // expected:
-            // 1) doc2 (1.0)
-            // 2) doc3 & doc4 both 2.0 → tie by original relevance: doc4(0.85) before doc3(0.75)
-            // 3) doc6(5),doc7(6),doc8(7),doc9(8),doc10(9)
-            // 4) missing last: doc1,doc5
-            HitGroup out = searcher.postProcessBySort(hitsToSort, sortJson, null, 10, 0);
-            assertThat(out.asList())
-                    .extracting(hit -> hit.getId().toString())
-                    .containsExactly(
-                            "doc2", "doc4", "doc3", "doc6", "doc7", "doc8", "doc9", "doc10", "doc1",
-                            "doc5");
-        }
-
-        /**
-         * Test that verifies sorting of results based on a single sort field with desc order
-         * and first missing policy.
-         */
-        @Test
-        void sort1FieldWithDescOrderAndFirstMissingPolicy() {
-            HitGroup hitsToSort = helpGenerateHitGroupWithOnlySortFieldValue0();
-            HybridSearcher searcher = new HybridSearcher();
-            String sortJson =
-                    "[{"
-                            + "\"field_name\":\"ignored\","
-                            + "\"order\":\"desc\","
-                            + "\"missing\":\"first\""
-                            + "}]";
-
-            HitGroup out = searcher.postProcessBySort(hitsToSort, sortJson, null, 10, 0);
-            assertThat(out.asList())
-                    .extracting(hit -> hit.getId().toString())
-                    .containsExactly(
-                            "doc1", // missing first  (–1e50, highest missing rel=0.55)
-                            "doc5", // missing second (–1e50, next missing rel=0.45)
-                            "doc10", // sort=9.0
-                            "doc9", // sort=8.0
-                            "doc8", // sort=7.0
-                            "doc7", // sort=6.0
-                            "doc6", // sort=5.0
-                            "doc4", // sort=2.0, tie-break on original rel=0.85 (before doc3)
-                            "doc3", // sort=2.0, tie-break rel=0.75
-                            "doc2" // sort=1.0
-                            );
-        }
-
-        /**
-         * Helper that builds 6 docs with two independent sort_field values:
-         *  - doc1/doc2 share field0=1.0 but doc2.field1=5.0 < doc1.field1=10.0
-         *  - doc3/doc4 share field0=2.0 but doc3.field1=3.0 < doc4.field1=7.0
-         *  - doc5/doc6 both missing (–1e50 → null)
-         */
-        HitGroup helpGenerateHitGroupWithTwoSortFieldValues() {
-            HitGroup hits = new HitGroup();
-            double[][] values = {
-                {1.0, 10.0},
-                {1.0, 5.0},
-                {2.0, 3.0},
-                {2.0, 7.0},
-                {-1e50, -1e50},
-                {-1e50, -1e50}
-            };
-            double[] relevances = {0.10, 0.20, 0.30, 0.40, 0.05, 0.06};
-            for (int i = 0; i < values.length; i++) {
-                FeatureData f = mock(FeatureData.class);
-                when(f.getDouble("sort_field_value_0")).thenReturn(values[i][0]);
-                when(f.getDouble("sort_field_value_1")).thenReturn(values[i][1]);
-                Hit h = new Hit("doc" + (i + 1), relevances[i]);
-                h.setField("matchfeatures", f);
-                hits.add(h);
-            }
-            return hits;
-        }
-
-        @Test
-        void sort2FieldsWithAscOrderAndLastMissingPolicy() {
-            HitGroup hitsToSort = helpGenerateHitGroupWithTwoSortFieldValues();
-            HybridSearcher searcher = new HybridSearcher();
-            // first sort_field_value_0 asc, missing last
-            // then sort_field_value_1 asc, missing last
-            String sortJson =
-                    "["
-                            + "{\"field_name\":\"ignored\",\"order\":\"asc\",\"missing\":\"last\"},"
-                            + "{\"field_name\":\"ignored\",\"order\":\"asc\",\"missing\":\"last\"}"
-                            + "]";
-
-            HitGroup out = searcher.postProcessBySort(hitsToSort, sortJson, null, 10, 0);
-            // Expected:
-            // 1) doc2 (1.0,5.0) before doc1 (1.0,10.0)
-            // 2) doc3 (2.0,3.0) before doc4 (2.0,7.0)
-            // 3) missing last: doc6 (rel=0.06) before doc5 (rel=0.05)
-            assertThat(out.asList())
-                    .extracting(hit -> hit.getId().toString())
-                    .containsExactly(
-                            "doc2", "doc1",
-                            "doc3", "doc4",
-                            "doc6", "doc5");
-        }
-
-        /**
-         * Helper that builds 4 docs with three sort_field values:
-         *  - docA/B/C all have field0=1.0 but differ on field1/field2
-         *  - docD missing all three
-         */
-        HitGroup helpGenerateHitGroupWithThreeSortFieldValues() {
-            HitGroup hits = new HitGroup();
-            String[] ids = {"docA", "docB", "docC", "docD"};
-            double[][] values = {
-                {1.0, 1.0, 3.0}, // docA
-                {1.0, 1.0, 2.0}, // docB
-                {1.0, 2.0, 1.0}, // docC
-                {-1e50, -1e50, -1e50} // docD missing all
-            };
-            double[] relevances = {0.40, 0.50, 0.60, 0.70};
-            for (int i = 0; i < ids.length; i++) {
-                FeatureData f = mock(FeatureData.class);
-                when(f.getDouble("sort_field_value_0")).thenReturn(values[i][0]);
-                when(f.getDouble("sort_field_value_1")).thenReturn(values[i][1]);
-                when(f.getDouble("sort_field_value_2")).thenReturn(values[i][2]);
-                Hit h = new Hit(ids[i], relevances[i]);
-                h.setField("matchfeatures", f);
-                hits.add(h);
-            }
-            return hits;
-        }
-
-        @Test
-        void sort3FieldsWithAscOrderAndFirstMissingPolicy() {
-            HitGroup hitsToSort = helpGenerateHitGroupWithThreeSortFieldValues();
-            HybridSearcher searcher = new HybridSearcher();
-            // all three ascending, missing first
-            String sortJson =
-                    "[{\"field_name\":\"ignored\",\"order\":\"asc\",\"missing\":\"first\"},"
-                        + "{\"field_name\":\"ignored\",\"order\":\"asc\",\"missing\":\"first\"},"
-                        + "{\"field_name\":\"ignored\",\"order\":\"asc\",\"missing\":\"first\"}]";
-
-            HitGroup out = searcher.postProcessBySort(hitsToSort, sortJson, null, 10, 0);
-            // Expected:
-            // 1) missing first: docD
-            // 2) among the rest, field0 ties=1.0 → use field1:
-            //      docA/B have field1=1.0 < docC.field1=2.0 → so docA & docB
-            //    then break tie on field2: docB(2.0) < docA(3.0)
-            // 3) then docC
-            assertThat(out.asList())
-                    .extracting(hit -> hit.getId().toString())
-                    .containsExactly("docD", "docB", "docA", "docC");
-        }
-
-        private Execution makeEmptyExec() {
-            Execution exec = mock(Execution.class);
-            when(exec.search(any(Query.class)))
-                    .thenAnswer(
-                            invocation -> {
-                                Query q = invocation.getArgument(0);
-                                return new Result(q, new HitGroup());
-                            });
-            return exec;
-        }
-
-        /**
-         * Test that verifies that postProcessBySort is called when sortBy is set in the query.
-         */
-        @Test
-        void whenSortByFields_set_postProcessBySortIsCalled() {
-            // 1) Create a Mockito spy on the real HybridSearcher
-            HybridSearcher spy = spy(new HybridSearcher());
-
-            // 2) Stub out createSubQuery (both overloads) so we never NPE inside it
-            doAnswer(inv -> inv.getArgument(0))
-                    .when(spy)
-                    .createSubQuery(any(Query.class), anyString(), anyString(), anyBoolean());
-            doAnswer(inv -> inv.getArgument(0))
-                    .when(spy)
-                    .createSubQuery(
-                            any(Query.class), anyString(), anyString(), anyBoolean(), anyString());
-
-            // 3) Stub extractTensorRankFeature to return:
-            //    • null for “mult_weights_global” or “add_weights_global”
-            //    • an empty Tensor for everything else (fields_to_rank_*)
-            doAnswer(
-                            inv -> {
-                                String name = inv.getArgument(1);
-                                if (name.contains("mult_weights_global")
-                                        || name.contains("add_weights_global")) {
-                                    return null;
-                                }
-                                // non-null so createSubQuery and co. won’t blow up
-                                return Tensor.from("tensor<float>()");
-                            })
-                    .when(spy)
-                    .extractTensorRankFeature(any(Query.class), anyString());
-
-            // 4) Stub postProcessBySort so it just returns an empty HitGroup
-            doReturn(new HitGroup())
-                    .when(spy)
-                    .postProcessBySort(any(HitGroup.class), anyString(), any(), anyInt(), anyInt());
-
-            // 5) Build a Query that triggers the sortBy branch
-            Query q = new Query("?q");
-            q.properties().set("hits", 1);
-            q.properties().set("offset", 0);
-            q.properties().set("marqo__hybrid.retrievalMethod", "lexical");
-            q.properties().set("marqo__hybrid.rankingMethod", "lexical");
-            q.properties()
-                    .set(
-                            "marqo__hybrid.sortBy.fields",
-                            "[{\"field_name\":\"foo\",\"order\":\"asc\",\"missing\":\"last\"}]");
-            // MUST set this to avoid the NPE you saw
-            q.properties().set("marqo__hybrid.sortBy.minSortCandidates", 10);
-
-            // 6) Call search()
-            spy.search(q, makeEmptyExec());
-
-            // 7) Verify that only postProcessBySort() ran
-            verify(spy, times(1))
-                    .postProcessBySort(any(HitGroup.class), anyString(), any(), anyInt(), anyInt());
-            verify(spy, never())
-                    .postProcessResults(any(), any(), any(), anyInt(), anyInt(), anyBoolean());
-        }
-
-        /**
-         * Test that verifies that neither postProcessBySort nor postProcessResults is
-         * called when there is no sortBy and no global‐modifier tensors in the query.
-         */
-        @Test
-        void whenNoSortByAndNoModifiers_noPostProcessingMethodsAreCalled() {
-            // 1) Spy on the real HybridSearcher
-            HybridSearcher spy = spy(new HybridSearcher());
-
-            // 2) Stub out both overloads of createSubQuery to bypass its internals
-            doAnswer(inv -> inv.getArgument(0))
-                    .when(spy)
-                    .createSubQuery(any(Query.class), anyString(), anyString(), anyBoolean());
-            doAnswer(inv -> inv.getArgument(0))
-                    .when(spy)
-                    .createSubQuery(
-                            any(Query.class), anyString(), anyString(), anyBoolean(), anyString());
-
-            // 3) Stub extractTensorRankFeature to always return null
-            //    (so both mult_weights_global and add_weights_global are 'absent')
-            doReturn(null).when(spy).extractTensorRankFeature(any(Query.class), anyString());
-
-            // 4) Also stub the two post‐processors so that, if they *did* get called,
-            //    they’d return an empty HitGroup instead of blowing up
-            doReturn(new HitGroup())
-                    .when(spy)
-                    .postProcessBySort(any(HitGroup.class), anyString(), any(), anyInt(), anyInt());
-            doReturn(new HitGroup())
-                    .when(spy)
-                    .postProcessResults(any(), any(), any(), anyInt(), anyInt(), anyBoolean());
-
-            // 5) Build a Query with *no* sortBy.fields and *no* modifier tensors
-            Query q = new Query("?q");
-            q.properties().set("hits", 1);
-            q.properties().set("offset", 0);
-            q.properties().set("marqo__hybrid.retrievalMethod", "lexical");
-            q.properties().set("marqo__hybrid.rankingMethod", "lexical");
-            // note: we do NOT set marqo__hybrid.sortBy.fields
-            // and extractTensorRankFeature will return null for all names
-
-            // 6) Execute
-            spy.search(q, makeEmptyExec());
-
-            // 7) Assert that neither branch ran
-            verify(spy, never())
-                    .postProcessBySort(any(HitGroup.class), anyString(), any(), anyInt(), anyInt());
-            verify(spy, never())
-                    .postProcessResults(any(), any(), any(), anyInt(), anyInt(), anyBoolean());
-        }
-
-        /*
-           Test that verifies that postProcessResults is called when only modifiers exist
-           (i.e., no sortBy.fields).
-        */
-        @Test
-        void whenOnlyModifiersExist_postProcessResultsIsCalled() {
-            HybridSearcher spy = spy(new HybridSearcher());
-            doAnswer(inv -> inv.getArgument(0))
-                    .when(spy)
-                    .createSubQuery(any(), anyString(), anyString(), anyBoolean());
-            // simulate “has a global mult modifier” but no sortBy
-            Tensor dummy = Tensor.from("tensor<float>(d0[1]):[1]");
-            doReturn(dummy)
-                    .when(spy)
-                    .extractTensorRankFeature(any(), contains("mult_weights_global"));
-            doReturn(null)
-                    .when(spy)
-                    .extractTensorRankFeature(any(), contains("add_weights_global"));
-            doReturn(new HitGroup())
-                    .when(spy)
-                    .postProcessResults(any(), any(), any(), anyInt(), anyInt(), anyBoolean());
-
-            Query q = new Query("?q");
-            q.properties().set("hits", 1);
-            q.properties().set("offset", 0);
-            q.properties().set("marqo__hybrid.retrievalMethod", "lexical");
-            q.properties().set("marqo__hybrid.rankingMethod", "lexical");
-            // no sortBy.fields
-
-            spy.search(q, makeEmptyExec());
-
-            verify(spy, times(1)).postProcessResults(any(), eq(q), any(), eq(1), eq(0), eq(false));
-            verify(spy, never()).postProcessBySort(any(), anyString(), any(), anyInt(), anyInt());
-        }
-    }
-
-    @Nested
-    class RelevanceCutoffTest {
-
-        @Nested
-        class ReadRelevanceCutoffParameterTest {
-
-            @Test
-            void shouldReturnNullWhenMethodIsNull() {
-                Query query = new Query("search/?query=test");
-                // Use reflection to call private method
-                Double result = callReadRelevanceCutoffParameter(query, null);
-                assertThat(result).isNull();
-            }
-
-            @Test
-            void shouldReturnRelativeScoreFactorWhenMethodIsRelativeMaxScore() {
-                Query query = new Query("search/?query=test");
-                query.properties()
-                        .set("marqo__hybrid.relevanceCutoff.parameters.relativeScoreFactor", 0.8);
-
-                Double result = callReadRelevanceCutoffParameter(query, "relative_max_score");
-                assertThat(result).isEqualTo(0.8);
-            }
-
-            @Test
-            void shouldThrowExceptionWhenRelativeScoreFactorIsMissing() {
-                Query query = new Query("search/?query=test");
-
-                RuntimeException exception =
-                        assertThrows(
-                                RuntimeException.class,
-                                () ->
-                                        callReadRelevanceCutoffParameter(
-                                                query, "relative_max_score"));
-                assertThat(exception.getMessage())
-                        .contains(
-                                "marqo__hybrid.relevanceCutoff.parameters.relativeScoreFactor is"
-                                        + " missing");
-            }
-
-            @Test
-            void shouldReturnMeanStdDevFactorWhenMethodIsMeanStdDev() {
-                Query query = new Query("search/?query=test");
-                query.properties()
-                        .set("marqo__hybrid.relevanceCutoff.parameters.meanStdDevFactor", 1.5);
-
-                Double result = callReadRelevanceCutoffParameter(query, "mean_std_dev");
-                assertThat(result).isEqualTo(1.5);
-            }
-
-            @Test
-            void shouldThrowExceptionWhenMeanStdDevFactorIsMissing() {
-                Query query = new Query("search/?query=test");
-
-                RuntimeException exception =
-                        assertThrows(
-                                RuntimeException.class,
-                                () -> callReadRelevanceCutoffParameter(query, "mean_std_dev"));
-                assertThat(exception.getMessage())
-                        .contains(
-                                "marqo__hybrid.relevanceCutoff.parameters.meanStdDevFactor is"
-                                        + " missing");
-            }
-
-            @Test
-            void shouldReturnNullWhenMethodIsGapDetection() {
-                Query query = new Query("search/?query=test");
-
-                Double result = callReadRelevanceCutoffParameter(query, "gap_detection");
-                assertThat(result).isNull();
-            }
-
-            @Test
-            void shouldThrowExceptionForUnknownMethod() {
-                Query query = new Query("search/?query=test");
-
-                RuntimeException exception =
-                        assertThrows(
-                                RuntimeException.class,
-                                () -> callReadRelevanceCutoffParameter(query, "unknown_method"));
-                assertThat(exception.getMessage())
-                        .contains("Unknown relevance cutoff method: unknown_method");
-            }
-
-            private Double callReadRelevanceCutoffParameter(Query query, String method) {
-                try {
-                    java.lang.reflect.Method readMethod =
-                            HybridSearcher.class.getDeclaredMethod(
-                                    "readRelevanceCutoffParameter", Query.class, String.class);
-                    readMethod.setAccessible(true);
-                    return (Double) readMethod.invoke(hybridSearcher, query, method);
-                } catch (Exception e) {
-                    if (e.getCause() instanceof RuntimeException) {
-                        throw (RuntimeException) e.getCause();
-                    }
-                    throw new RuntimeException(e);
-                }
-            }
-        }
-
-        @Nested
-        class DetectCutoffCountTest {
-
-            @Test
-            void shouldReturnZeroForEmptyHitGroup() {
-                HitGroup emptyHits = new HitGroup();
-
-                Integer result = callDetectCutoffCount(emptyHits, "gap_detection", null, false);
-                assertThat(result).isEqualTo(0);
-            }
-
-            @Nested
-            class GapDetectionTest {
-
-                @Test
-                void shouldFindGapInScores() {
-                    HitGroup hits = createHitGroupWithScores(1.0, 0.9, 0.8, 0.3, 0.2, 0.1);
-
-                    Integer result = callDetectCutoffCount(hits, "gap_detection", null, false);
-                    // Gap between 0.8 and 0.3 is largest (0.5), so cutoff at index 3
-                    assertThat(result).isEqualTo(3);
-                }
-
-                @Test
-                void shouldHandleUniformGaps() {
-                    HitGroup hits = createHitGroupWithScores(1.0, 0.9, 0.8, 0.7, 0.6, 0.5);
-
-                    Integer result = callDetectCutoffCount(hits, "gap_detection", null, false);
-                    // With uniform gaps (0.1 each), algorithm returns 3 based on actual behavior
-                    assertThat(result).isEqualTo(3);
-                }
-
-                @Test
-                void shouldHandleSingleHit() {
-                    HitGroup hits = createHitGroupWithScores(1.0);
-
-                    Integer result = callDetectCutoffCount(hits, "gap_detection", null, false);
-                    assertThat(result).isEqualTo(1);
-                }
-
-                @Test
-                void shouldFindEarliestGap() {
-                    HitGroup hits = createHitGroupWithScores(1.0, 0.5, 0.4, 0.1);
-
-                    Integer result = callDetectCutoffCount(hits, "gap_detection", null, false);
-                    // Gap between 1.0 and 0.5 is 0.5, between 0.4 and 0.1 is 0.3
-                    // Largest gap is at index 1
-                    assertThat(result).isEqualTo(1);
-                }
-            }
-
-            @Nested
-            class MeanStdDevTest {
-
-                @Test
-                void shouldCountHitsAboveThreshold() {
-                    // Scores: 1.0, 0.8, 0.6, 0.4, 0.2
-                    // Mean = 0.6, StdDev ≈ 0.283
-                    // With factor 1.0: threshold = 0.6 + 0.283 = 0.883
-                    // Only score 1.0 is above threshold
-                    HitGroup hits = createHitGroupWithScores(1.0, 0.8, 0.6, 0.4, 0.2);
-
-                    Integer result = callDetectCutoffCount(hits, "mean_std_dev", 1.0, false);
-                    assertThat(result).isEqualTo(1);
-                }
-
-                @Test
-                void shouldCountHitsAboveMeanPlusStdDev() {
-                    // Scores: [1.0, 0.9, 0.8, 0.7, 0.6]
-                    // Mean = 0.8, StdDev ≈ 0.1265
-                    // With factor 0.1: threshold = 0.8 + (0.1265 * 0.1) = 0.81265
-                    // Hits above 0.81265: 1.0, 0.9 (0.8 is below threshold)
-                    HitGroup hits = createHitGroupWithScores(1.0, 0.9, 0.8, 0.7, 0.6);
-
-                    Integer result = callDetectCutoffCount(hits, "mean_std_dev", 0.1, false);
-                    assertThat(result).isEqualTo(2);
-                }
-
-                @Test
-                void shouldCountZeroWhenNoneAboveThreshold() {
-                    HitGroup hits = createHitGroupWithScores(0.5, 0.4, 0.3, 0.2, 0.1);
-
-                    Integer result = callDetectCutoffCount(hits, "mean_std_dev", 3.0, false);
-                    assertThat(result).isEqualTo(0);
-                }
-            }
-
-            @Nested
-            class RelativeMaxScoreTest {
-
-                @Test
-                void shouldCountHitsAboveRelativeThreshold() {
-                    HitGroup hits = createHitGroupWithScores(1.0, 0.9, 0.7, 0.5, 0.3);
-
-                    // With factor 0.8: threshold = 1.0 * 0.8 = 0.8
-                    // Hits above 0.8: 1.0, 0.9
-                    Integer result = callDetectCutoffCount(hits, "relative_max_score", 0.8, false);
-                    assertThat(result).isEqualTo(2);
-                }
-
-                @Test
-                void shouldCountAllHitsWithLowThreshold() {
-                    HitGroup hits = createHitGroupWithScores(1.0, 0.8, 0.6, 0.4, 0.2);
-
-                    Integer result = callDetectCutoffCount(hits, "relative_max_score", 0.1, false);
-                    assertThat(result).isEqualTo(5);
-                }
-
-                @Test
-                void shouldCountOnlyTopHitWithHighThreshold() {
-                    HitGroup hits = createHitGroupWithScores(1.0, 0.8, 0.6, 0.4, 0.2);
-
-                    Integer result = callDetectCutoffCount(hits, "relative_max_score", 1.0, false);
-                    assertThat(result).isEqualTo(1);
-                }
-            }
-
-            @Test
-            void shouldThrowExceptionForUnknownMethod() {
-                HitGroup hits = createHitGroupWithScores(1.0, 0.5);
-
-                RuntimeException exception =
-                        assertThrows(
-                                RuntimeException.class,
-                                () -> callDetectCutoffCount(hits, "unknown_method", 0.5, false));
-                assertThat(exception.getMessage())
-                        .contains("Unknown relevance cutoff method: unknown_method");
-            }
-
-            private HitGroup createHitGroupWithScores(double... scores) {
-                HitGroup hits = new HitGroup();
-                for (int i = 0; i < scores.length; i++) {
-                    hits.add(new Hit("index:test/0/doc" + i, scores[i]));
-                }
-                return hits;
-            }
-
-            private Integer callDetectCutoffCount(
-                    HitGroup hits, String method, Double parameter, boolean verbose) {
-                try {
-                    java.lang.reflect.Method detectMethod =
-                            HybridSearcher.class.getDeclaredMethod(
-                                    "detectCutoffCount",
-                                    HitGroup.class,
-                                    String.class,
-                                    Double.class,
-                                    boolean.class);
-                    detectMethod.setAccessible(true);
-                    return (Integer)
-                            detectMethod.invoke(hybridSearcher, hits, method, parameter, verbose);
-                } catch (Exception e) {
-                    if (e.getCause() instanceof RuntimeException) {
-                        throw (RuntimeException) e.getCause();
-                    }
-                    throw new RuntimeException(e);
-                }
-            }
-        }
-    }
 }
